requirements.txt
python-telegram-bot==11.1.0
requests==2.19.1
schedule==0.5.0